+++ src/main/java/pl/plajer/murdermystery/api/StatsStorage.java
@@ -103,7 +103,8 @@ public class StatsStorage {
    */
   public enum StatisticType {
     CONTRIBUTION_DETECTIVE("contribdetective", true), CONTRIBUTION_MURDERER("contribmurderer", true), DEATHS("deaths", true), GAMES_PLAYED("gamesplayed", true), HIGHEST_SCORE("highestscore", true),
-    KILLS("kills", true), LOSES("loses", true), WINS("wins", true), LOCAL_GOLD("gold", false), LOCAL_KILLS("local_kills", false), LOCAL_SCORE("local_score", false);
+    KILLS("kills", true), LOSES("loses", true), WINS("wins", true), LOCAL_CURRENT_PRAY("local_pray", false), LOCAL_GOLD("gold", false), LOCAL_KILLS("local_kills", false),
+    LOCAL_PRAISES("local_praises", false), LOCAL_SCORE("local_score", false);
 
     private String name;
     private boolean persistent;