+++ src/main/java/pl/plajer/murdermystery/arena/ArenaManager.java
@@ -117,7 +117,11 @@ public class ArenaManager {
         p.setFlying(true);
         User user = UserManager.getUser(p.getUniqueId());
         user.setSpectator(true);
-        user.setStat(StatsStorage.StatisticType.LOCAL_GOLD, 0);
+        for(StatsStorage.StatisticType stat : StatsStorage.StatisticType.values()) {
+          if(!stat.isPersistent()) {
+            user.setStat(stat, 0);
+          }
+        }
         p.addPotionEffect(new PotionEffect(PotionEffectType.NIGHT_VISION, Integer.MAX_VALUE, 0));
         ArenaUtils.hidePlayer(p, arena);
 
@@ -251,9 +255,11 @@ public class ArenaManager {
       if (!plugin.isBungeeActivated() && plugin.isInventoryManagerEnabled()) {
         InventoryUtils.loadInventory(plugin, p);
       }
-      user.setStat(StatsStorage.StatisticType.LOCAL_GOLD, 0);
-      user.setStat(StatsStorage.StatisticType.LOCAL_SCORE, 0);
-      user.setStat(StatsStorage.StatisticType.LOCAL_KILLS, 0);
+      for(StatsStorage.StatisticType stat : StatsStorage.StatisticType.values()) {
+        if(!stat.isPersistent()) {
+          user.setStat(stat, 0);
+        }
+      }
     } catch (Exception ex) {
       new ReportedException(plugin, ex);
     }